% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso3166}
\alias{iso3166}
\title{Country list}
\format{
A data frame with 250 rows 8 variables:
\itemize{
\item{country_name: }
\item{continent: }
\item{region: }
\item{iso2c: }
\item{iso3c: }
\item{iso3n: 3 digits number (as character)}
\item{x: Coordinates for displaying as tabularmap}
\item{y: Coordinates for displaying as tabularmap}
}
}
\usage{
iso3166
}
\description{
A data frame include ISO-3166 codes.
}
\keyword{datasets}
