% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jpn77}
\alias{jpn77}
\title{Japan 7x7 grid dataset}
\format{A data frame with 47 rows 11 variables:
\itemize{
\item{jis_code: jis code}
\item{prefecture: prefecture names}
\item{capital: capital name for prefecture}
\item{region: region}
\item{major_island: }
\item{prefecture_en: }
\item{capital_en: }
\item{region_en: }
\item{major_island_en: }
\item{x: Coordinates for displaying as tabularmap}
\item{y: Coordinates for displaying as tabularmap}
}}
\usage{
jpn77
}
\description{
Prefectures dataset.
}
\keyword{datasets}
