% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularmap.R
\name{scale_fill_jpregion}
\alias{scale_fill_jpregion}
\title{Coloring the tabularmaps by region in Japan}
\usage{
scale_fill_jpregion(lang, ...)
}
\arguments{
\item{lang}{Select whether the region variable is Japanese (\code{jp}) or English (\code{en}).}

\item{...}{all other arguments passed on to \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}}
}
\description{
Custom ggplot2 scale for tabulamap.
}
