% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularmap.R
\name{tabularmap}
\alias{tabularmap}
\title{Create Tile-Grid}
\usage{
tabularmap(data, fill = 1, label, ...)
}
\arguments{
\item{data}{data.frame. Contain x, y variables used as coordinates.}

\item{fill}{Fill colour variable.}

\item{label}{Label variable.}

\item{...}{All other arguments passed on to \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} include label \code{family}.}
}
\description{
A ggplot2-based tabularmap that places a coordinated dataset in a rectangle.
}
\examples{
library(ggplot2)
tabularmap(jpn77, label = prefecture, size = 3)
tabularmap(jpn77, fill = region, label = prefecture, size = 3)
tabularmap(jpn77, fill = region, label = prefecture, size = 3) +
  theme_tabularmap() +
  scale_fill_jpregion(lang = "en")
tabularmap(data.frame(
  x = rep(c(1,2,3), each = 3),
  y = rep(c(1,2,3), times = 3),
  fill = seq.int(9),
  label = letters[seq.int(9)]),
  fill = fill, label = label)
}
