% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularmap.R
\name{theme_tabularmap}
\alias{theme_tabularmap}
\title{Tabularmap theme}
\usage{
theme_tabularmap(...)
}
\arguments{
\item{...}{all other arguments passed on to \code{\link[ggplot2:theme_minimal]{ggplot2::theme_minimal()}}}
}
\description{
Custom ggplot2 theme for tabulamap.
}
