% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tky23}
\alias{tky23}
\title{Special wards of Tokyo}
\format{
A data frame with 23 rows 5 variables:
\itemize{
\item no: Identifical number
\item ward: Name
\item ward_kanji: Names in Kanji
\item x, y: Coordinates for displaying as tabularmap
}
}
\usage{
tky23
}
\description{
Tokyo 23 wards dataset.
}
\keyword{datasets}
