% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jpn77}
\alias{jpn77}
\title{Japan 7x7 grid dataset}
\format{
A data frame with 47 rows 8 variables:
\itemize{
\item{jis_code: jis code}
\item{prefecture: prefecture names}
\item{region: region}
\item{major_island: }
\item{prefecture_kanji: }
\item{region_kanji: }
\item{x: Coordinates for displaying as tabularmap}
\item{y: Coordinates for displaying as tabularmap}
}
}
\usage{
jpn77
}
\description{
Prefectures dataset.
}
\keyword{datasets}
